Convert ContentStatistics.
